+++ src/gateways/Gateway.php
@@ -32,6 +32,7 @@ use craft\commerce\stripe\errors\CustomerException;
 use craft\commerce\stripe\errors\PaymentSourceException;
 use craft\commerce\stripe\events\BuildGatewayRequestEvent;
 use craft\commerce\stripe\events\CreateInvoiceEvent;
+use craft\commerce\stripe\events\Receive3dsPaymentEvent;
 use craft\commerce\stripe\events\ReceiveWebhookEvent;
 use craft\commerce\stripe\models\forms\CancelSubscription;
 use craft\commerce\stripe\models\forms\Payment;
@@ -138,6 +139,29 @@ class Gateway extends BaseGateway
      * ```
      */
     const EVENT_RECEIVE_WEBHOOK = 'receiveWebhook';
+
+    /**
+     * @event Receive3dsPaymentEvent The event that is triggered when a successful 3ds payment is received.
+     *
+     * Plugins get a chance to do something whenever a successful 3D Secure payment is received.
+     *
+     * ```php
+     * use craft\commerce\Plugin as Commerce;
+     * use craft\commerce\stripe\events\Receive3dsPaymentEvent;
+     * use craft\commerce\stripe\gateways\Gateway as StripeGateway;
+     * use yii\base\Event;
+     *
+     * Event::on(StripeGateway::class, StripeGateway::EVENT_RECEIVE_3DS_PAYMENT, function(Receive3dsPaymentEvent $e) {
+     *     $order = $e->transaction->getOrder();
+     *     $orderStatus = Commerce::getInstance()->getOrderStatuses()->getOrderStatusByHandle('paid');
+     *     if ($order && $paidStatus && $order->orderStatusId !== $paidStatus->id && $order->getIsPaid()) {
+     *         $order->orderStatusId = $paidStatus->id;
+     *         Craft::$app->getElements()->saveElement($order);
+     *     }
+     * });
+     * ```
+     */
+    const EVENT_RECEIVE_3DS_PAYMENT = 'receive3dsPayment';
 
     /**
      * string The Stripe API version to use.
@@ -1082,7 +1106,7 @@ class Gateway extends BaseGateway
             return;
         }
 
-        $childTransaction = Commerce::getInstance()->getTransactions()->createTransaction(null,     $transaction);
+        $childTransaction = Commerce::getInstance()->getTransactions()->createTransaction(null, $transaction);
         $childTransaction->reference = $data['id'];
 
         try {
@@ -1124,6 +1148,15 @@ class Gateway extends BaseGateway
             }
 
             Commerce::getInstance()->getTransactions()->saveTransaction($childTransaction);
+
+            if ($childTransaction->status === TransactionRecord::STATUS_SUCCESS) {
+                if ($this->hasEventHandlers(self::EVENT_RECEIVE_3DS_PAYMENT)) {
+                    $this->trigger(self::EVENT_RECEIVE_3DS_PAYMENT, new Receive3dsPaymentEvent([
+                        'transaction' => $childTransaction
+                    ]));
+                }
+
+            }
         } catch (\Exception $exception) {
             Craft::error('Could not process webhook '.$data['id'].': '.$exception->getMessage(), 'stripe');
             $childTransaction->status = TransactionRecord::STATUS_FAILED;